% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_edits.R
\name{get_all_edits}
\alias{get_all_edits}
\title{Title}
\usage{
get_all_edits(users, lang = "pt", sys_sleep = FALSE, sleep_seconds = 5)
}
\arguments{
\item{sleep_seconds}{}
}
\value{

}
\description{
Title
}
