% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_edits.R
\name{user_edits_json}
\alias{user_edits_json}
\title{Title}
\usage{
user_edits_json(id_user, lang)
}
\arguments{
\item{lang}{}

\item{sleep_seconds}{}
}
\value{



}
\description{
Title
}
\examples{
Title

}
